Added two checks for database initialization and placeholder prints
In database.py the init_db checks for the presence of the database
path.  If it doesn't exist, print statements are placeholders for
calling Nicolet5's database population code.

After init_db, a check for an existing user named 'admin' is
performed.  If it returns nothing, a new user is created.