Added integration of database and data table
Currently, the User table is hardcoded to be displayed in the data table on the main page, however, we will change that to the MasterTable once populated and eventually it will be automatically updated by the dropdown experiment type selection.

Also added a custom JSONEncoder mixin (from https://github.com/dacb/viscount/blob/a1965fcf0537a64fcaa60f31e9b00fb09afc4ce0/viscount/utils.py) so that the results of the sqlalchemy queries can be jsonified. A work around for setting the default JSON encoding of the AppenderQuery object had to be set to None.